Use 'vanilla' websocket components
Back in the day that OFMeet was created, the websocket support in Openfire was poor.
Nowadays, that's a lot better. When we use the Openfire-provided websocket support,
we can remove a lot of custom code in OFMeet (notably, the custom websocket
implementation, the Strophe modifications to make use of the custom websockets, and
the client modifications to make use of he Strophe modifications).

By reducing the amount of customizations, it should be less hard to update third
party components.