Use SASL for authentication
The proprietary websocket implementation that was replaced by a 'vanilla' variant that's
provided by Openfire contained code that simulated some sort of SSO between the OFMeet
applications (Jitsi Meet, Candy, etc). By removing the properietary websocket
implementation, that authentication mechanism was removed.

This commit introduces a new authentication mechanism, that is based on a SASL mechanism.
When the user accesses OFmeet, the application server (Jetty) will trigger authentication.
This commit ensures that upon successful authentication, an access token is created, which,
obtained through the configuration servlet, is used by the web applications.