Replaced customized Jitsi Videobridge plugin code
Previously, OFMeet contained Jitsi Videobridge code (and its Openfire plugin) that was modified for OFMeet. This fork is undesireable for many of the common reasons that forks are undesirable.

This commit replaces the custom Jitsi Videobridge code in OFMeet with a set of libraries, taken directly from Jitsi*. Any customizations are made in classes that are specific to OFMeet.

The state of the code after this commit is applied should be somewhat stable/usable, but is expected to be followed up by more commits. Your mileage might vary and no warrentee on anything...

* The Jitsi libraries included in this commit are in reality a local build, that include jitsi code + three to-be-merged PRs. This is expected to be replaced by proper Jitsi-provided libraries before we release this code.

Changes in this commit
- jitsi-plugin made standard. Deleted copy of jitsi code. Custom modifications moved here:
-- Openfire properties to system settings (to configure jitsi)
-- autorecord should become jitsi videobridge feature: https://github.com/jitsi/jitsi-videobridge/issues/344
- jicofo moved from (modified) jitsiplugin and moved to this class (still uses copy of jitsi code)

Not done (yet?)
- MAX/MIN_PORT_DEFAULT_VALUE: older OFMeet moved the range from 5000-6000 (jitsi's default) to 50000-60000.
- BundleActivator jitsiActivator: removed / no longer needed?