Work around singleton OSGi
Our plugin combines two Jitsi components, that each make use of OSGi: the Jitsi Videobridge Openfire plugin and the Jitsi focus component.

Jitsi's OSGi implementation is a static implementation, ment to be initialized exactly once. This leaves the second component that we start in a configuration limbo.

To work around this issue, this commit changes management of the OSGi lifecycle of the Jitsi focus component. Instead of being managed by the Jitsi OSGi implementation, management now occurs within our code. As long as no other components try to access the OSGi configuration for the focus component through the OSGi implemention of Jitsi, this should be safe.