New build process for OFMeet
Instead of maintaining copies of Jitsi artifacts in our repository, these artifacts should be downloaded on demand. This allows us to stop wondering if we put in the right artifacts (and looks at the project definition of third-party projects instead), and also reduces the sheer size of the project.

With this change, the ant buildscript:
- clones the source repository for Jitsi components
- uses Maven to download all of their dependencies.
- uses Maven to build the code
- combines these results with our own sources to construct the OFMeet plugin.

The buildscript introduces two new properties, in which the GIT tag numbers of the desired state of the Jitsi components can be configured. When that version of the code has previously been checked out, no new checkout is made (which speeds up the build considerably).

Apache Ant, plus some related libraries, have been added to this repository. This simplifies execution, as the integration of Ant and those libraries takes some consideration. The build scripts have been modified to make use of this included Ant.

A prerequisite for the build script now is Maven, which must be installed on the system that executes the build. Additionally, a M2_HOME system property must point to its installation directory.