Jitsi Meet component should not depend on root context deployment.
The upstream Jitsi Meet application depends on deployment on the root context of a URL (for
example: http://meet.example.org). For OFMeet, it is desirable (needed even) to use a non-root
context (for example: http://example.org/ofmeet).

The original OFMeet code had custom modifications to the Jitsi Meet application (introducing
a query parameter). This commit replaces those modifications with a solution that no longer
uses a query parameter. This solution is expected to be accepted upstream, which would make
'our' copy of Jitsi Meet be an unmodified copy of the original again.

Specifically, this commit:
- rolls back custom modifications
- applies changes that have also been supplied in an upsteam PR

Upstream PR: https://github.com/jitsi/jitsi-meet/pull/1202